% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_and_split_chimeric_reads.R
\name{chimeric_mapping_validation}
\alias{chimeric_mapping_validation}
\title{Compare chimeric segments and reads using alignment data}
\usage{
chimeric_mapping_validation(
  BAM_path = NULL,
  CHIMERIC_INFO = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  THREADS = 1,
  OVERLAP_THRESHOLD = 0.8,
  silent = FALSE
)
}
\arguments{
\item{BAM_path}{character. "/path/to/bam.sorted.bam" BAM must have been produced with both raw reads and chimeric splitted reads.}

\item{CHIMERIC_INFO}{character. "/path/to/_chimericSegments.info.txt" from 'run_blast_and_split_reads()'.}

\item{OUTPUT_DIR}{character. "/path/to/output/"}

\item{OUTPUT_SUFFIX}{character.}

\item{THREADS}{numeric.}

\item{OVERLAP_THRESHOLD}{numeric. Percentage of overlap between the union of chimeric segment and read mappings to consider the segment as chimeric read split as correct.}

\item{silent}{boolean. If TRUE will print various messages.}
}
\value{
data.frame - _chimeric_banList.txt

data.frame - _in_silico_chimerics.log.txt

data.frame - _mapping_chimerics.log.txt
}
\description{
Summary: Validate the split of chimeric reads into sub-segments using mapping information. Exclude in-silico chimeric.

PART-1: Ligation Chimerics

1. Map chimeric segments and chimeric reads using Minimap2 ('run_minimap2()').

2. Extract the mapped segments (M: mapped, S: soft-clipped, H: Hard-clipped) in both primary and supplementary mappings.

3. Convert the read/segment mapped into GRanges objects.

4. For each read, compare the mapped segments to the read mapping. Assumption: the chimeric segments align at the same position as the read.

5. Calculate the overlap between the mapped segments and the union of the mapped segment/read. Preferred over a simple % of intersection to avoid this case:


Segment:    |===|


Read:   |===========|


Intersection shows 100% but union <50%.

6. For each segment / read overlap, select the one with the highest overlap.

7. Segment which have < "OVERLAP_THRESHOLD" with the read mapping portions are discarded.

This severs as a validation of the chimeric split. Will exclude both false negative and positive chimeric events.

PART-2: In-silico Chimerics

8. Detect segments of a same read mapping to the same genomic position (overlap >"OVERLAP_THRESHOLD").

9. If they display opposite orientations = in-silico chimerics.

10. For each in-silico pairs, select the longest segment length as the correct one.

Final:

Report the segment_ids which do not display enough overlap with the read mapping.

Report the shortest segments among in-silico chimeric reads.

This function do not discard any segment / read. It will flag them as problematic but it is up-to the user to decide if they should be removed.
I recommend using Dorado + duplex-tools to split and use in-silico chimerics for duplex. This should minimise the in-silico chimeric left-overs.
}
