% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFiles.R
\name{combineFiles}
\alias{combineFiles}
\title{Combine FASTQ and Barcode Files}
\usage{
combineFiles(
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  VALIDATE_CHIMERIC = TRUE,
  UMI_TYPE = "NONE",
  THREADS = 4,
  silent = TRUE,
  STRAND_SPECIFIC = NULL
)
}
\arguments{
\item{OUTPUT_DIR}{character. Output directory of the other functions.}

\item{OUTPUT_SUFFIX}{character. Sample ID.}

\item{VALIDATE_CHIMERIC}{boolean. If chimeric mapping validation was run, will return the combined files.}

\item{UMI_TYPE}{character. "NONE", "MONOMER" or "TRIMER".}

\item{THREADS}{numeric.}

\item{silent}{boolean. Should some extra info be shown ?}

\item{STRAND_SPECIFIC}{boolean, were the strand specific files generated?}
}
\description{
Combine results from the parallel processes before demultiplexing
}
