% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umi_tools.R
\name{convert_isoquant_UMItools2}
\alias{convert_isoquant_UMItools2}
\title{convert_isoquant_UMItools2}
\usage{
convert_isoquant_UMItools2(
  ISOQUANT_BAM_FOLDER = NULL,
  silent = TRUE,
  mode.bam = "NONE",
  UPDATED_READ_ASSIGN = NULL,
  OUTPUT_SUFFIX = NULL,
  OUTPUT_DIR = NULL,
  THREADS = 4
)
}
\arguments{
\item{ISOQUANT_BAM_FOLDER}{character. "/path/to/the/aux/" from isoquant (contains the BAM Files)}

\item{silent}{boolean. Suppress some of the messages.}

\item{mode.bam}{"10x" or "NONE", "MONOMER", or "TRIMER"}

\item{UPDATED_READ_ASSIGN}{character. "/path/to/updated_read_assignments.txt" from update_read_assignments()}

\item{OUTPUT_SUFFIX}{character. Sample id to label the outputs.}

\item{OUTPUT_DIR}{character. "/path/to/the/output/"}

\item{THREADS}{numeric. Number of parallel sessions.}
}
\description{
Run a python script to add new SAM tags to a BAM. Works by crossing the BAM files with a txt file containing the new tags in columns (2 characters columns) and the read_id column (first col)
}
