% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch_barcodes_demultiplexing.R
\name{determine_strand_orientation}
\alias{determine_strand_orientation}
\title{determine_strand_orientation}
\usage{
determine_strand_orientation(
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  dT_SEQ = "AAGCAGTGGTATCAACGCAGAGTACTTTTTTTTTTTTTTTTTTTTT",
  THREADS = 1,
  WINDOW = 200,
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (.gz) or not.}

\item{OUTPUT_DIR}{character. /path/to/output_dir/.}

\item{OUTPUT_SUFFIX}{character. Sample id.}

\item{dT_SEQ}{character. Sequence of the dT (no UMI / Ns allowed)}

\item{THREADS}{numeric. Number of threads to run bbmap.}

\item{WINDOW}{numeric. Up/down window in the reads to look for the dT}

\item{silent}{boolean. Display messages or not.}
}
\value{
*_detected_dT.txt object
}
\description{
Detect the position of the dT in the read. Useful to reorient the reads afterwards.
}
