% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBarcode.R
\name{extractBarcodes}
\alias{extractBarcodes}
\title{Extract barcodes from VSEARCH results}
\usage{
extractBarcodes(
  VSEARCH_RES_PATH = NULL,
  silent = FALSE,
  WHITELIST_PATH = NULL,
  SELECTED_BC_IDs = NULL,
  THREADS = 4,
  MIN_READS = 100,
  PRIMER_SEQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL
)
}
\arguments{
\item{VSEARCH_RES_PATH}{character. "/path/to/the/input/vsearch_results.txt".}

\item{silent}{boolean. Display messages / stats.}

\item{WHITELIST_PATH}{character. "/path/to/the/whitelist_index.fasta".}

\item{SELECTED_BC_IDs}{character. Vector of the barcode IDs to consider. Ex. c("BC1", "BC2", ...).}

\item{THREADS}{numeric. Thread number to calculate the distances.}

\item{MIN_READS}{numeric. Output only indexes supported with at least MIN_READS}

\item{PRIMER_SEQ}{character.}

\item{OUTPUT_DIR}{character. "/path/to/output/".}

\item{OUTPUT_SUFFIX}{character.}
}
\value{
dataframe associating the most likely index to each read.
}
\description{
Extract the barcode sequences from the VSEARCH results ('run_vsearch()').
Compare them to a index whitelist. Sequencing / PCR errors are corrected by taking the closest matching barcode sequence.
Discard reads with barcodes too distant from the whitelist (>MAX_LEVENSTEIN). Given the current barcode design, it is not recommended to push >=3.
}
