% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch_barcodes_demultiplexing.R
\name{filter_trim_demultiplex_OLD}
\alias{filter_trim_demultiplex_OLD}
\title{filter Trim and demultiplex reads}
\usage{
filter_trim_demultiplex_OLD(
  BARCODES_PATH = NULL,
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  THREADS = 1,
  N_READS = 10000,
  UMI_TYPE = "NONE",
  STRAND_SPECIFIC = FALSE,
  WINDOW = NULL,
  BANLIST = NULL,
  TRIM = TRUE,
  VSEARCH_BARCODE_RESULTS = NULL,
  silent = FALSE
)
}
\arguments{
\item{BARCODES_PATH}{data.frame. Output from 'extractBarcodes()'.}

\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (.gz) or not.}

\item{OUTPUT_DIR}{character. /path/to/output_dir/.}

\item{OUTPUT_SUFFIX}{character. Sample id.}

\item{THREADS}{numeric. Number of threads to run bbmap.}

\item{N_READS}{numeric. For the demultiplexing, iterate through chunks of N_READS (see ShortRead::FastqStreamer).}

\item{UMI_TYPE, }{character. Accepts "NONE", "MONOMER", "TRIMER" or default to no UMI.}

\item{WINDOW}{numeric. Run vsearch only on the read start/end windows. Must be the same as WINDOW in vsearch_run.}

\item{BANLIST}{character. "/path/to/banlist_dataframe.txt". Created with 'chimeric_mapping_validation()'}

\item{TRIM}{boolean. (Beta) If TRUE, will search for the PRIMER_SEQ and trim anything from it to the read ends. Currently the trimming will leave CCG (TSO) or T(30) (oligo-dT) at the end of the reads.}

\item{VSEARCH_BARCODE_RESULTS}{character. If trimming, get the barcode positions from vsearch.}

\item{silent}{boolean. Display messages or not.}
}
\value{
Write demultiplexed FASTQ files to OUTPUT_DIR
}
\description{
Take the output from 'extractBarcodes()' and remove the chimeric indexes combinations and optionally select specific barcodes from the whitelist.
Create individual files containing the read IDs associated to a specific barcode
}
\details{
THIS FUNCTION WILL LIKELY BE SPLIT IN MULTIPLE PARTS - BECAME TOO BIG.
}
