% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barcode_list.R
\name{generate_index_seq}
\alias{generate_index_seq}
\title{generate_index_seq_V1}
\usage{
generate_index_seq(
  size = 13,
  dist = 4,
  metric = "seqlev",
  threads = 24,
  heuristic = "conway",
  outpath = "Desktop/",
  context = "CGACGCTCTTCCGATCTNNNNNNNNNNNNNAAGCAGTGGTATCAACGCA",
  nanopore_bc = "/opt/ont/guppy/data/barcoding/barcodes_masked.fasta",
  nanopore_kit_prefix = "NB"
)
}
\arguments{
\item{size}{numeric. Barcode length (max 14).}

\item{dist}{numeric. Maximum number of substitutions / distance. See DNABarcodes::create.dnabarcodes.}

\item{metric}{character. "seqlev" or "hamming"}

\item{threads}{numeric. DNABarcodes::create.dnabarcodes threads. Only useful with heuristic = ashlock.}

\item{heuristic}{character. Tried 'conway' (fast) and ashlock (slow).}

\item{outpath}{character. /path/to/results/}

\item{context}{character. DNA barcode final context marked with Ns "CGACGCTCTTCCGATCTNNNNNNNNNNNNNAAGCAGTGGTATCAACGCA".}

\item{nanopore_bc}{character. "/path/to/the/output/native/barcoding/barcode/sequence.fasta}

\item{nanopore_kit_prefix}{character. Example: "NB"}
}
\description{
Create a set of random barcodes using DNABarcodes. Can be very slow if long barcodes are required (>24hrs).
This function is tailored for this particular application and was not extensively benchmarked with other parameters than those proposed.
}
\details{
1. Run DNABarcodes::create.dnabarcodes using the defined parameters.
==> Idealy I aimed at 12bp bc / seqlev=4 but this function kept returning 320 bc (60 short from full-plate).
==> Went for 13bp instead.
2. Filter out sequences with triplets (including those with the barcode edges, ex. T .. A)
3. Filter out sequences with >2 homodimers (including those at the barcode edge, ex. T .. A)
4. Filter out GC content >65% and <35%.
5. Exclude barcodes too close from the nanopore_bc / adapters (lv dist <10).
6. Sample randomly 384 barcodes 10000 times and calculate their levenstein distance.
7. Get the barcodes list with the smallest mean levenstein distance to maximise the chances during demultiplexing.
8. Return the barcodes as a fasta file.
}
