% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_UMI.R
\name{get_and_correct_UMI_trimer}
\alias{get_and_correct_UMI_trimer}
\title{Extract and Correct UMI sequences from reads (trimers)}
\usage{
get_and_correct_UMI_trimer(
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  TRIMERS = c("TTC", "GGT", "AAA", "CCG"),
  dT_SEQ = "AAGCAGTGGTATCAACGCAGAGTNNNNNNNNNNNNNNNNNNNNNNNNCTTTTTTTTTT",
  THREADS = 1,
  WINDOW = 200,
  WHITELIST = "CONTENT_AWARE",
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (expect '.gz') or not.}

\item{OUTPUT_DIR}{character. "/path/to/output/"}

\item{OUTPUT_SUFFIX}{character. sample id.}

\item{TRIMERS}{character vector.  c("TTC", "GGT", "AAA", "CCG").}

\item{dT_SEQ}{character. Should be set to "AAGCAGTGGTATCAACGCAGAGTNNNNNNNNNNNNNNNNNNNNNNNNCTTTTTTTTTT".}

\item{THREADS}{numeric. Number of threads. Will remove 1 for safety.}

\item{WINDOW}{numeric. Detection windows for the UMI at the read start/ends.}

\item{WHITELIST}{"ALL" or "CONTENT_AWARE". If ALL, use all the trimer combination for the whitelist. If CONTENT_AWARE, use all the unique perfect trimer combinations detected during the first pass.}

\item{silent}{boolean. If TRUE will print various messages.}
}
\description{
Detect the UMI (trimer) position in the read (+/- 70%). Correct for indels / mutations - return the UMI associated to each read (+/- 85% efficacy)
}
