% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_UMI.R
\name{get_UMI}
\alias{get_UMI}
\title{Extract and Correct UMI sequences from reads}
\usage{
get_UMI(
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  dT_SEQ = "AAGCAGTGGTATCAACGCAGAGTACNNNNNNNNATACTGACGCTTT",
  THREADS = 1,
  MAX_GAP_LENGTH = 2,
  MAX_GAPS_N = 2,
  WINDOW = 200,
  MAX_MISMATCH_N = 6,
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (expect '.gz') or not.}

\item{OUTPUT_DIR}{character. "/path/to/output/"}

\item{OUTPUT_SUFFIX}{character. sample id.}

\item{dT_SEQ}{character. Should be set to "AAGCAGTGGTATCAACGCAGAGTCTTTNNNNNNNNNNNNNNNNNNNNNNNNTTTTTTTTTT".}

\item{THREADS}{numeric. Number of threads. Will remove 1 for safety.}

\item{MAX_GAP_LENGTH}{numeric. Max gap length in detected dT_SEQ.}

\item{MAX_GAPS_N}{numeric. Max number of gaps allowed in detected dT_SEQ.}

\item{MAX_MISMATCH_N}{numeric. Max number of mismatches allowed in detected dT_SEQ.}

\item{silent}{boolean. If TRUE will print various messages.}
}
\description{
Detect the UMI (trimer) position in the read.

NB: Do not modify the minimal seq identity - it was chosen to maximise the recovery of UMI while excluding oligo-dT without UMI.
Will report <0.2% of matching reads in samples without oligo-dT-UMI.
}
