% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umi_tools.R
\name{isoquant_count_matrix_inconsistent}
\alias{isoquant_count_matrix_inconsistent}
\title{Isoquant Update Transcript Count Matrix}
\usage{
isoquant_count_matrix_inconsistent(
  UPDATED_READ_ASSIGN = NULL,
  read_bc = "detected_barcodes.txt",
  OUTPUT_SUFFIX = NULL,
  OUTPUT_DIR = NULL,
  silent = TRUE
)
}
\arguments{
\item{UPDATED_READ_ASSIGN}{character. "/path/to/updated_read_assignments.txt" from update_read_assignments()}

\item{read_bc}{character. "/path/to/FSnanopore_detected_barcodes.txt" (contains columns named "read_id", "index_unique")}

\item{silent}{boolean. TRUE or FALSE.}

\item{output_dir}{character. Output directory of the other functions.}

\item{output_suffix}{character. Sample ID.}
}
\description{
Using the updated read assignment file, create new grouped count matrixes for gene stringent (XR) and transcripts stringents but full/partial spliced (XI).
}
