% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment_and_BAM_manipulations.R
\name{load_bam}
\alias{load_bam}
\title{load_bam}
\usage{
load_bam(
  BAM_path = NULL,
  tidy_chimeric = TRUE,
  clip_flag_width = 300,
  THREADS = 1,
  FIELDS = c("qname", "flag", "rname", "strand", "pos", "qwidth", "mapq", "cigar")
)
}
\arguments{
\item{BAM_path}{character. "/path/to/the/input/file.sorted.bam". Requires an indexed file in the same folder.}

\item{clip_flag_width}{numeric. Flag as "putative chimeric" reads with a soft-clipped / hard-clipped value superior to 'soft_clip_flag_width' (left/right).}

\item{FIELDS}{character vector. BAM fields to read. Expect the following for chimeric detection: c("qname","flag", "rname","strand","pos", "qwidth", "mapq", "cigar")}

\item{getSoftClipping}{boolean. If TRUE, will report the right / left clipping (soft & hard) start/end/width in the read. Useful to spot putative chimeric reads and compare read mapping segments.}
}
\value{
BAM file to data.frame.
}
\description{
Load a BAM file to R.
If getSoftClipping = TRUE, will also report the position on the read of left/right-clippings and mapping.
}
