% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareFASTQ.R
\name{prepareFASTQ}
\alias{prepareFASTQ}
\title{Split FASTQ file by line and filter by length}
\usage{
prepareFASTQ(
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  N_CHUNKS = NULL,
  THREADS = NULL,
  OUTPUT_SUFFIX = NULL,
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (expect '.gz') or not.}

\item{OUTPUT_DIR}{character. "/path/to/output/"}

\item{N_CHUNKS}{numeric. In how many pieces should the FASTQ be split ?}

\item{THREADS}{numeric.}

\item{OUTPUT_SUFFIX}{character. sample id.}

\item{silent}{boolean. If TRUE will print various messages.}
}
\description{
Split a FASTQ file in equivalent chunks of reads based on the number of threads.
}
