% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isoquant.R
\name{random_sampling}
\alias{random_sampling}
\title{random FASTQ sampling}
\usage{
random_sampling(
  fq.input = NULL,
  n_reads = NULL,
  OUTPUT_SUFFIX = NULL,
  OUTPUT_DIR = NULL
)
}
\arguments{
\item{fq.input}{character. FASTQ path (gz or not - detected with grepl).}

\item{n_reads}{numeric. Number of reads to downsample.}

\item{OUTPUT_SUFFIX}{character. Sample ID.}

\item{OUTPUT_DIR}{character. Output directory of the other functions.}
}
\description{
Randomly sample a defined number of reads from a FASTQ file. Use seqkit sample
}
