% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_barcode_list.R
\name{randomBC}
\alias{randomBC}
\title{generate_index_seq_V2}
\usage{
randomBC(
  iterations = 100,
  n = 1e+05,
  l = 20,
  n_bc = 96,
  nucl = c("A", "T", "C", "G")
)
}
\arguments{
\item{iterations}{numeric. Barcode length (max 14).}

\item{n}{numeric. Maximum number of substitutions / distance. See DNABarcodes::create.dnabarcodes.}

\item{l}{character. "seqlev" or "hamming"}

\item{n_bc}{numeric. How many BC to pick?}

\item{nucl}{character. Tried 'conway' (fast) and ashlock (slow).}
}
\value{
data.frame with the following columns:
barcodes sequence
minimal levenstein distance of this barcode to the other in that set
set (=iteration)
min.lv.set minimal levenstein distance of that set
}
\description{
Create a set of random barcodes.
}
\examples{
bc18 <- randomBC(iterations = 500, n = 200000, l = 18, nucl = c("A", "T", "C", "G"))
bc19 <- randomBC(iterations = 500, n = 200000, l = 19, nucl = c("A", "T", "C", "G"))
bc20 <- randomBC(iterations = 500, n = 200000, l = 20, nucl = c("A", "T", "C", "G"))
bc21 <- randomBC(iterations = 500, n = 200000, l = 21, nucl = c("A", "T", "C", "G"))
bc22 <- randomBC(iterations = 500, n = 200000, l = 22, nucl = c("A", "T", "C", "G"))
# Select bc20 with the least number of BC harboring a minimal lv distance to the other barcode (=6)
index <- bc20 \%>\%
  group_by(it_n) \%>\%
  summarise(bc_lv6 = sum(min_dist == 6, na.rm = T))

final.bc <- bc20 \%>\%
  filter(it_n == index$it_n[min(index$bc_lv6)]) \%>\%
  filter(min_dist >= 7) \%>\%
  head(384)

hist(lv_dist_to_df(final.bc$row, final.bc$row)$min_dist)
hist(str_count(final.bc$row, "G|C")/20)

final.bc <- final.bc$row
names(final.bc) <- paste0("BC", 1:384)

ShortRead::writeFasta(Biostrings::DNAStringSet(final.bc), "/home/vincent.hahaut/Desktop/IOB/6_Nanopore-FS/FSnanoporeR/inst/extdata/whitelist.bc20.fa")

}
