% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_and_split_chimeric_reads.R
\name{run_blast}
\alias{run_blast}
\title{Run BLAST using a single target sequence}
\usage{
run_blast(
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  PCR_ADAPTER_REV = "AAGCAGTGGTATCAACGCAGAGT",
  PCR_ADAPTER_FWD = "AAGCAGTGGTATCAACGCAGAGT",
  THREADS = 1,
  IDENTITY = 65,
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (expect '.gz') or not.}

\item{OUTPUT_DIR}{character. "/path/to/output/"}

\item{OUTPUT_SUFFIX}{character.}

\item{PCR_ADAPTER_REV}{character. Sequence searched with BLASTn to define the read orientation and chimerism.}

\item{PCR_ADAPTER_FWD}{character. Sequence searched with BLASTn to define the read orientation and chimerism.}

\item{THREADS}{numeric.}

\item{IDENTITY}{numeric. Percentage of identity between query and "PCR_ADAPTER" to be considered as a valid "PCR_ADAPTER" hit.}

\item{silent}{boolean. If TRUE will print various messages.}
}
\value{
data.frame - _blast_search.txt
}
\description{
Convert FASTQ into FASTA. Create a blast 5 database using the adapter sequence and map the reads against it.
}
