% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_isoquant.R
\name{run_isoquant}
\alias{run_isoquant}
\title{random FASTQ sampling}
\usage{
run_isoquant(
  downsample_n_reads = "",
  output_dir = NULL,
  output_suffix = NULL,
  transcript_counts_type = "unique_only",
  premapping = TRUE,
  n_cores = NULL,
  isoquant_fa = NULL,
  isoquant_gtf = NULL,
  max_memory = NULL,
  minimap2_binaries = NULL,
  minimap2_ref = NULL,
  minimap2_bed = NULL
)
}
\arguments{
\item{downsample_n_reads}{numeric or "". If numeric, will resample X reads from each cell.}

\item{output_dir}{character. Output directory of the other functions.}

\item{output_suffix}{character. Sample ID.}

\item{transcript_counts_type}{character. Accepts "unique_only" or "with_inconsistent". See isoquant.}

\item{premapping}{boolean. If FALSE, let isoquant map your samples. If TRUE, map the samples in parallel prior to isoquant.}

\item{isoquant_fa}{character. /path/to/reference/genome.fa.}

\item{isoquant_gtf}{character. /path/to/genome/annotation.gtf.}

\item{max_memory}{see run_minimap2.}

\item{minimap2_binaries}{see run_minimap2.}

\item{minimap2_ref}{see run_minimap2.}

\item{minimap2_bed}{see run_minimap2.}
}
\description{
Run isoquant on the samples.
}
