% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Alignment_and_BAM_manipulations.R
\name{run_minimap2}
\alias{run_minimap2}
\title{run_minimap2}
\usage{
run_minimap2(
  FASTQ = NULL,
  OUTPUT_SUFFIX = NULL,
  MINIMAP2_Binaries = NULL,
  MINIMAP2_REF = NULL,
  MINIMAP2_BED = NULL,
  OUTPUT_DIR = "./",
  CHIMERIC_ON = FALSE,
  STRAND_SPECIFIC = FALSE,
  ISOQUANT_MODE = FALSE,
  THREADS = 1
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (expect '.gz') or not. Accept also a vector of FASTQ file paths.}

\item{OUTPUT_SUFFIX}{character. Sample prefix id.}

\item{MINIMAP2_Binaries}{character. "/path/to/minimap2" binaries}

\item{MINIMAP2_REF}{character. "/path/to/minimap2_reference.fa" FASTA or .mni. Using FASTA will significantly slow down the process.}

\item{MINIMAP2_BED}{character. "/path/to/minimap2_gene_reference.bed" prepared using paftools.js}

\item{OUTPUT_DIR}{character. "/path/to/the/output/".}

\item{CHIMERIC_ON}{boolean. If TRUE will select sam flag '260' else '2308'.}

\item{STRAND_SPECIFIC}{boolean. Should it be run only considering the forward strand (-uf).}

\item{ISOQUANT_MODE}{boolean. If TRUE, use the -Y flag to mark secondary alignments as soft-cliping.}

\item{THREADS}{numeric. Number of threads.}
}
\value{
Creates a sorted BAM file at paste0(OUTPUT_DIR, "/", OUTPUT_SUFFIX, "_sorted.bam").
}
\description{
Map FASTQ file using Minimap2 and filter out unmapped reads.

Requires minimap2 path. Requires samtools installed.
}
\details{
https://github.com/lh3/minimap2/blob/master/misc/README.md
}
