% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umi_tools.R
\name{run_umi_tools}
\alias{run_umi_tools}
\title{run_umi_tools}
\usage{
run_umi_tools(
  BAM = NULL,
  OUTPUT_SUFFIX = NULL,
  OUTPUT_DIR = NULL,
  silent = TRUE,
  UMI_TOOLS_BIN = "umi_tools"
)
}
\arguments{
\item{BAM}{character. "/path/to/the/file_isoquant_umi.bam" converted using convert_isoquant_UMItools.}

\item{OUTPUT_SUFFIX}{character. Sample id to label the outputs.}

\item{OUTPUT_DIR}{character. "/path/to/the/output_directory"}

\item{UMI_TOOLS_BIN, }{character. "Path/to/umi_tools" binaries. If not set, default to "umi_tools".}
}
\description{
Run umi_tools on one or multiple bam files using first the assigned gene tag and then the transcript gene tag from isoquant.
}
