% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_vsearch.R
\name{run_vsearch}
\alias{run_vsearch}
\title{run_vsearch but on the read ends}
\usage{
run_vsearch(
  FASTQ = NULL,
  OUTPUT_DIR = "./",
  FORWARD_SEQ = NULL,
  REVERSE_SEQ = NULL,
  SEQ_IDENTITY = 0.75,
  OUTPUT_SUFFIX = "bc",
  MIN_SEQ_LENGTH = 30,
  WINDOW = 200,
  THREADS = 1,
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. /path/to/the/input/fastq.gz. Gziped (.gz) or not (detected with grepl).}

\item{OUTPUT_DIR}{character. /path/to/the/output_directory/}

\item{FORWARD_SEQ}{character. Sequence of the forward sequence to search for (5==>3).}

\item{REVERSE_SEQ}{character. Sequence of the reverse sequence to search for (5==>3).}

\item{SEQ_IDENTITY}{numeric. Minimal identity required between the barcode sequence and the query.}

\item{OUTPUT_SUFFIX}{character. Sample id to label the outputs.}

\item{MIN_SEQ_LENGTH}{numeric. Minimal match length between the barcode sequence and the query.}

\item{WINDOW}{numeric. Run vsearch only on the read start/end windows.}

\item{THREADS}{numeric. Number of threads to run vsearch}

\item{silent}{boolean. Display messages or not.}
}
\value{
Write the results to paste0(OUTPUT_DIR, "vsearch_", SAMPLE_ID, ".txt").
}
\description{
Search for a specific set of barcode sequences in FASTQ reads. Vsearch is preferred over BLAST here because of the way it handles Ns and the output format.
Altogether this provides a slightly better performance than BLAST at retrieving the primers.
The difference between run_vsearch and run_vsearch2 is that the later first extract the WINDOW bp (ex 200bp) from the read end/start and then look for the barcode in that specific region while the other does it accross the entire read body.
Probably feasible with BLAST but quite simple with VSEARCH once chimeric reads have been split.
Warning: Will only find one hit per read which prevents it from detecting chimeric molecules other than those with two different barcodes!

Requires vsearch installed and in your path (https://github.com/torognes/vsearch).
}
\details{
Vsearch fields: https://www.drive5.com/usearch/manual/userfields.html
}
