% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_FASTQ_management.R
\name{split_and_write_fastq}
\alias{split_and_write_fastq}
\title{Write Chimeric Segments}
\usage{
split_and_write_fastq(
  FASTQ = NULL,
  CHIMERIC_SEGMENTS_PATH = NULL,
  OUTPUT_PATH = NULL,
  CHIMERIC_READ_IDs = NULL,
  OUTPUT_SUFFIX = NULL
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (.gz) or not.}

\item{CHIMERIC_SEGMENTS_PATH}{character. Path to the blast search object created by 'detect_and_split_chimeric()'}

\item{OUTPUT_PATH}{character. "/path/to/the/output/folder/"}

\item{CHIMERIC_READ_IDs}{character. Path to the chimeric read ids.}

\item{OUTPUT_SUFFIX}{character. Sample ID}
}
\value{
Three fastq files

.chimeric_reads.fq.gz: Chimeric reads before splitting

.non_chimeric_reads.fq.gz: Non Chimeric reads

.chimeric_segments.fq.gz: Chimeric reads split into segments
}
\description{
REQUIRE: seqkit
}
\details{
Take a FASTQ fle and divide into non_chimeric reads and chimeric ones. Chimeric ones are further spitted into segments.
The read_id will be modified, keeping only the prefix and appending:

.chimeric_reads.fq.gz ==> prefix + ""

.non_chimeric_reads.fq.gz ==> prefix + "_chimeric=FALSE_1"

.chimeric_segments.fq.gz ==> prefix + "_chimeric=TRUE_" + read_segment_id (unique)
}
