% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_and_split_chimeric_reads.R
\name{split_chimeric_reads}
\alias{split_chimeric_reads}
\title{Detect and Split Chimeric Reads}
\usage{
split_chimeric_reads(
  FASTQ = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL,
  PCR_ADAPTER_REV = "",
  PCR_ADAPTER_FWD = "",
  CHIMERIC_DETECTION_WINDOW = 200,
  IDENTITY = 65,
  silent = FALSE
)
}
\arguments{
\item{FASTQ}{character. "/path/to/the/input/fastq.gz". Gziped (expect '.gz') or not.}

\item{OUTPUT_DIR}{character. "/path/to/output/"}

\item{OUTPUT_SUFFIX}{character.}

\item{PCR_ADAPTER_REV}{character. Sequence searched with BLASTn to define the read orientation and chimerism.}

\item{PCR_ADAPTER_FWD}{character. Sequence searched with BLASTn to define the read orientation and chimerism.}

\item{CHIMERIC_DETECTION_WINDOW}{numeric. Consider a read as chimeric if it has a "PCR_ADAPTER" sequence located >'CHIMERIC_DETECTION_WINDOW' bp from the 5' or 3' edge.}

\item{IDENTITY}{numeric. Percentage of identity between query and "PCR_ADAPTER" to be considered as a valid "PCR_ADAPTER" hit.}

\item{silent}{boolean. If TRUE will print various messages.}
}
\value{
data.frame - _chimericSegments.info.txt

FASTQ.gz #1 - _chimeric_reads.fq.gz

FASTQ.gz #2 - _chimeric_segments.fq.gz

FASTQ.gz #3 - _non_chimeric_reads.fq.gz
}
\description{
Summary: Identity chimeric reads and split them into segments.

1. Run BLASTn-Short looking for any "PCR_ADAPTER" in the read with an identity > "IDENTITY".

2. Chimeric reads are defined as having at least a "PCR_ADAPTER" located > "CHIMERIC_DETECTION_WINDOW" base-pairs of the 5' or 3' end.

3. Ideal reads have a 5' PCR_ADAPTER(+) and a 3' PCR_ADAPTER(-). Parse the BLAST output and define split the chimeric reads into segments according to the "PCR_ADAPTER" orientations such as:


Read:
(5')...<==|==>...<==|==>...(3')


Segments:


(+)==>...<==(-)


(+)==>...(read_end)


(read_start)...<==(-)


4. Filter out segments <100 bp.

5. Add a 50bp at the "PCR_ADAPTER" 3' end to encompass the cell barcode #1 if present. Currently hard-coded to 50bp as I except than large window values may be problematic for downstream analysis. This approach will likely extract a few bases from neighboring chimeric segments. However, minimap2 will align them as soft-clipped sequences given their short size and they are unlikely to interefere with the rest of the analysis.

6. Correct the read length to avoid going <0 bp or >read_length.

7. Save the BLAST result table.

8. Use 'split_and_write_fastq()' to create three FASTQ.gz files: non-chimeric reads, chimeric reads and chimeric segments (=splitted reads).

It should be noted that any method solely based on the detection of the "PCR_ADAPTER" to split chimeric reads will miss the following chimeric case:


(+) ==>... | ... <== (-)


Where two reads are ligated in opposite orientations and lack a central "PCR_ADAPTER". Although rare, these cases are handled by 'chimeric_mapping_validation()'.

NOTE: require blast installed. Expect BLAST version compatible with blast database version 5. BLAST is used instead of VSEARCH because it can return multiple similar hits per read.
}
