% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED_FASTQ_management.R
\name{stream_FASTQ_stats}
\alias{stream_FASTQ_stats}
\title{Get FASTQ stats}
\usage{
stream_FASTQ_stats(fq = NULL)
}
\arguments{
\item{fq}{character. FASTQ path}
}
\description{
Get read_id, read_length, read_average_bp_qual (phred). Stream over a fastq file by 10000 reads to avoid overloading the memory
}
