% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umi_tools.R
\name{tidy_umi_tools}
\alias{tidy_umi_tools}
\title{tidy_umi_tools}
\usage{
tidy_umi_tools(
  UMI_TOOLS_RESULTS_FOLDER = NULL,
  OUTPUT_DIR = NULL,
  OUTPUT_SUFFIX = NULL
)
}
\arguments{
\item{UMI_TOOLS_RESULTS_FOLDER}{character. Sample id to label the outputs.}

\item{OUTPUT_DIR}{character. "/path/to/the/output_directory"}

\item{OUTPUT_SUFFIX}{character. Sample id to label the outputs.}
}
\value{
OUTPUT_DIR/OUTPUT_SUFFIX_isoquant_umitools_counts.gene.tsv.gz and OUTPUT_DIR/OUTPUT_SUFFIX_isoquant_umitools_counts.transcripts.tsv.gz files
}
\description{
Create a matrix of counts using umi_Tools results
}
