% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umi_tools.R
\name{update_read_assignments}
\alias{update_read_assignments}
\title{update_read_assignments}
\usage{
update_read_assignments(
  ISOQUANT_READ_ASSIGNMENTS = NULL,
  mode.bam = NULL,
  OUTPUT_SUFFIX = NULL,
  OUTPUT_DIR = NULL,
  silent = FALSE
)
}
\arguments{
\item{ISOQUANT_READ_ASSIGNMENTS}{character. "/path/to/the/input/*.read_assignments.tsv" from isoquant.}

\item{mode.bam}{"NONE", "MONOMER", or "TRIMER"}

\item{OUTPUT_SUFFIX}{character. Sample id to label the outputs.}

\item{OUTPUT_DIR}{character. "/path/to/the/output/"}

\item{silent}{boolean. Display messages or not.}
}
\description{
Update the read assignment file from Isoquant to add new SAM tags
}
\details{
NB: There is definitely a faster way to do it - I will explore options later.
}
